+++ src/models/api/responses/nft_info.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
@@ -7,6 +8,6 @@ pub struct NftInfo {
     pub owner: Option<String>,
     pub description: Option<String>,
     pub image_url: Option<String>,
-    pub list_price: Option<i64>,
-    pub top_offer: Option<i64>,
+    pub list_price: Option<BigDecimal>,
+    pub top_offer: Option<BigDecimal>,
 }
