+++ src/models/db/listing.rs
@@ -1,3 +1,4 @@
+use bigdecimal::BigDecimal;
 use chrono::NaiveDateTime;
 use serde::{Deserialize, Serialize};
 
@@ -11,7 +12,7 @@ pub struct Listing {
     pub collection_id: Option<String>,
     pub nft_id: Option<String>,
     pub nonce: Option<String>,
-    pub price: Option<i64>,
+    pub price: Option<BigDecimal>,
     pub price_str: Option<String>,
     pub seller: Option<String>,
     pub tx_index: Option<i64>,
